move putting of attrib to state tree to StateTreeStore
