+++ sovrin_node/persistence/StateTreeStore.py
@@ -7,6 +7,8 @@ from sovrin_common.txn import TXN_TYPE, \
     ISSUER_KEY, GET_ISSUER_KEY, REF, TRUSTEE, TGB, IDENTITY_TXN_TYPES, \
     CONFIG_TXN_TYPES, POOL_UPGRADE, ACTION, START, CANCEL, SCHEDULE, \
     NODE_UPGRADE, COMPLETE, FAIL, HASH, ENC, RAW, NONCE
+import json
+from hashlib import sha256
 
 # TODO: think about encapsulating State in it,
 # instead of direct accessing to it in node
@@ -27,23 +29,43 @@ class StateTreeStore:
         """
 
         {
-            NYM: self._addNym,
-            ATTRIB: self._addAttrib,
-            SCHEMA: self._addSchema,
+            NYM:        self._addNym,
+            ATTRIB:     self._addAttrib,
+            SCHEMA:     self._addSchema,
             ISSUER_KEY: self._addIssuerKey
         }[txn[TXN_TYPE]](txn)
 
 
-
-
     def _addNym(self, txn):
+        assert txn[TXN_TYPE] == NYM
         raise NotImplementedError
 
     def _addAttrib(self, txn):
-        raise NotImplementedError
+        assert txn[TXN_TYPE] == ATTRIB
+
+        def parse(txn):
+            raw = txn.get(RAW)
+            if raw:
+                data = json.loads(raw)
+                key, value = data.items()[0]  # only one attr per txn, yes
+                return key, value
+            enc = txn.get(ENC)
+            if enc:
+                return (enc, enc)
+            hash = txn.get(HASH)
+            if hash:
+                return hash, None
+            raise ValueError("One of 'raw', 'enc', 'hash' "
+                             "fields of ATTR must present")
+
+        rawKey, value = parse(txn)
+        key = sha256(rawKey.encode()).hexdigest()
+        self.state.set(key, value)
 
     def _addSchema(self, txn):
+        assert txn[TXN_TYPE] == SCHEMA
         raise NotImplementedError
 
     def _addIssuerKey(self, txn):
+        assert txn[TXN_TYPE] == ISSUER_KEY
         raise NotImplementedError
